+++ bindgen/lib.rs
@@ -171,7 +171,6 @@ pub enum Formatter {
     None,
     /// Use `rustfmt` to format the bindings.
     Rustfmt,
-    #[cfg(feature = "prettyplease")]
     /// Use `prettyplease` to format the bindings.
     Prettyplease,
 }
@@ -189,7 +188,6 @@ impl FromStr for Formatter {
         match s {
             "none" => Ok(Self::None),
             "rustfmt" => Ok(Self::Rustfmt),
-            #[cfg(feature = "prettyplease")]
             "prettyplease" => Ok(Self::Prettyplease),
             _ => Err(format!("`{}` is not a valid formatter", s)),
         }
@@ -201,7 +199,6 @@ impl std::fmt::Display for Formatter {
         let s = match self {
             Self::None => "none",
             Self::Rustfmt => "rustfmt",
-            #[cfg(feature = "prettyplease")]
             Self::Prettyplease => "prettyplease",
         };
 
@@ -966,8 +963,6 @@ impl Bindings {
 
         match self.options.formatter {
             Formatter::None => return Ok(tokens.to_string()),
-
-            #[cfg(feature = "prettyplease")]
             Formatter::Prettyplease => {
                 return Ok(prettyplease::unparse(&syn::parse_quote!(#tokens)));
             }